MAINT: stats: move _broadcast_arrays etc to appropriate file
